[routing] move world graph to its own dir
