+++ routing/world_graph/single_vehicle_world_graph.hpp
@@ -10,7 +10,8 @@
 #include "routing/route.hpp"
 #include "routing/segment.hpp"
 #include "routing/transit_info.hpp"
-#include "routing/world_graph.hpp"
+
+#include "routing/world_graph/world_graph.hpp"
 
 #include "routing_common/num_mwm_id.hpp"
